requirements.txt
python-dotenv
pandas
numpy
scikit-learn
ipykernel
certifi
pymongo[srv]==3.11
pymongo
scipy
PyYAML
# -e .